update RemoveModule [ClassController]

+++ Controllers/ClassController.cs
@@ -70,11 +70,11 @@ namespace kroniiapi.Controllers
         {
             (int totalRecord, IEnumerable<Class> classList) = await _classService.GetClassList(paginationParameter);
 
-            // foreach (Class c in classList)
-            // {
-            //     c.Trainer = await _trainerService.GetTrainerById(c.TrainerId);
-            //     c.Admin = await _adminService.GetAdminById(c.AdminId);
-            // }
+            foreach (Class c in classList)
+            {
+                //c.Trainer = await _trainerService.GetTrainerById(c.TrainerId);
+                c.Admin = await _adminService.GetAdminById(c.AdminId);
+            }
             IEnumerable<ClassResponse> classListDto = _mapper.Map<IEnumerable<ClassResponse>>(classList);
             if (totalRecord == 0)
             {
@@ -108,9 +108,9 @@ namespace kroniiapi.Controllers
         /// <returns>200: Update done / 404: Class or request not found / 409: Class or request deactivated</returns>
         [HttpPut("request/{deleteClassRequestId:int}")]
         [Authorize(Policy = "ClassPut")]
-        public async Task<ActionResult> ConfirmDeleteClassRequest([FromBody] ConfirmDeleteClassInput confirmDeleteClassInput, int deleteClassRequestId)
+        public async Task<ActionResult> ConfirmDeleteClassRequest([FromBody] ConfirmDeleteClassInput confirmDeleteClassInput,int deleteClassRequestId)
         {
-            int status = await _classService.UpdateDeletedClass(confirmDeleteClassInput, deleteClassRequestId);
+            int status = await _classService.UpdateDeletedClass(confirmDeleteClassInput,deleteClassRequestId);
             if (status == -1)
             {
                 return NotFound(new ResponseDTO(404, "Class or request not found"));
@@ -164,7 +164,6 @@ namespace kroniiapi.Controllers
             var cdr = _mapper.Map<ClassDetailResponse>(s);
             return Ok(cdr);
         }
-
         /// <summary>
         /// Get the detail information of a class 
         /// </summary>
@@ -596,8 +595,7 @@ namespace kroniiapi.Controllers
         [HttpGet("trainer/{id:int}")]
         public async Task<ActionResult<PaginationResponse<IEnumerable<TrainerClassListResponse>>>> GetClassListByTrainerId(int id, [FromQuery] PaginationParameter paginationParameter)
         {
-            if (!_trainerService.CheckTrainerExist(id))
-            {
+            if (!_trainerService.CheckTrainerExist(id)) {
                 return NotFound(new ResponseDTO(404, "Trainer not found"));
             }
             (int totalRecord, IEnumerable<Class> classList) = await _classService.GetClassListByTrainerId(id, paginationParameter);
@@ -632,14 +630,11 @@ namespace kroniiapi.Controllers
         public async Task<ActionResult> RemoveModule(AssignModuleInput assignModuleInput)
         {
             var classInfor = await _classService.GetClassByClassID(assignModuleInput.ClassId);
-            if (classInfor == null)
+            if(classInfor == null)
             {
                 return NotFound(new ResponseDTO(404, "Class is not exist"));
             }
-            // if(classInfor.TrainerId != assignModuleInput.TrainerId)
-            // {
-            //     return BadRequest(new ResponseDTO(409, "Trainer are not in this class"));
-            // }
+
             var moduleMarkState = await _markService.GetMarkByModuleId(assignModuleInput.ModuleId, null, null);
             if (moduleMarkState.Count() != 0)
             {
@@ -652,12 +647,18 @@ namespace kroniiapi.Controllers
                 return NotFound(new ResponseDTO(404, "Trainee in this class has certificate with this module"));
             }
 
+            var classModuleInfor = await _classService.GetClassModule(assignModuleInput.ClassId, assignModuleInput.ModuleId);
+            if(classModuleInfor.TrainerId != assignModuleInput.TrainerId)
+            {
+                return BadRequest(new ResponseDTO(409, "This trainer dont teach this class"));
+            }
+
             int removeStatus = await _classService.RemoveModuleFromClass(assignModuleInput.ClassId, assignModuleInput.ModuleId);
-            if (removeStatus == -1)
+            if(removeStatus == -1)
             {
                 return BadRequest(new ResponseDTO(409, "Class does not have this module"));
             }
-            else if (removeStatus == 1)
+            else if(removeStatus == 1)
             {
                 return Ok(new ResponseDTO(200, "Deleted!"));
             }
